+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * replace logError with logBrowser, adjust compliance set loading background ([5e0925b](https://github.com/ColdPBC/cold-monorepo/commit/5e0925b582b1e8a2b7707f01153ce5a1ec84b25c))
 
 
+
 ### Features
 
 * add supplier removal functionality for materials ([7787e8a](https://github.com/ColdPBC/cold-monorepo/commit/7787e8aaecb3ebcba7915b0ba0d44c035b8220cd))